app/order/requirements.txt
Flask==2.2.3
pika==1.3.0
psycopg2-binary==2.9.5
prometheus-client==0.16.0
